+++ core/execution/liquidation/engine.go
@@ -24,13 +24,14 @@ import (
 	"code.vegaprotocol.io/vega/core/events"
 	"code.vegaprotocol.io/vega/core/execution/common"
 	"code.vegaprotocol.io/vega/core/idgeneration"
+	"code.vegaprotocol.io/vega/core/positions"
 	"code.vegaprotocol.io/vega/core/types"
 	vegacontext "code.vegaprotocol.io/vega/libs/context"
 	"code.vegaprotocol.io/vega/libs/crypto"
 	"code.vegaprotocol.io/vega/libs/num"
 )
 
-//go:generate go run github.com/golang/mock/mockgen -destination mocks/mocks.go -package mocks code.vegaprotocol.io/vega/core/execution/liquidation Book,MarketLiquidity,IDGen
+//go:generate go run github.com/golang/mock/mockgen -destination mocks/mocks.go -package mocks code.vegaprotocol.io/vega/core/execution/liquidation Book,MarketLiquidity,IDGen,Positions,Settlement
 
 type Book interface {
 	GetVolumeAtPrice(price *num.Uint, side types.Side) uint64
@@ -44,6 +45,15 @@ type IDGen interface {
 	NextID() string
 }
 
+type Positions interface {
+	RegisterOrder(ctx context.Context, order *types.Order) *positions.MarketPosition
+	Update(ctx context.Context, trade *types.Trade, passiveOrder, aggressiveOrder *types.Order) []events.MarketPosition
+}
+
+type Settlement interface {
+	AddTrade(trade *types.Trade)
+}
+
 type Engine struct {
 	// settings, orderbook, network pos data
 	cfg      *types.LiquidationStrategy
@@ -55,6 +65,8 @@ type Engine struct {
 	nextStep time.Time
 	tSvc     common.TimeService
 	ml       MarketLiquidity
+	position Positions
+	settle   Settlement
 	stopped  bool
 }
 
@@ -89,20 +101,22 @@ func GetLegacyStrat() *types.LiquidationStrategy {
 	return legacyStrat.DeepClone()
 }
 
-func New(cfg *types.LiquidationStrategy, mktID string, broker common.Broker, book Book, as common.AuctionState, tSvc common.TimeService, ml MarketLiquidity) *Engine {
+func New(cfg *types.LiquidationStrategy, mktID string, broker common.Broker, book Book, as common.AuctionState, tSvc common.TimeService, ml MarketLiquidity, pe Positions, se Settlement) *Engine {
 	// NOTE: This can be removed after protocol upgrade
 	if cfg == nil {
 		cfg = legacyStrat.DeepClone()
 	}
 	return &Engine{
-		cfg:    cfg,
-		broker: broker,
-		mID:    mktID,
-		book:   book,
-		as:     as,
-		tSvc:   tSvc,
-		ml:     ml,
-		pos:    &Pos{},
+		cfg:      cfg,
+		broker:   broker,
+		mID:      mktID,
+		book:     book,
+		as:       as,
+		tSvc:     tSvc,
+		ml:       ml,
+		position: pe,
+		settle:   se,
+		pos:      &Pos{},
 	}
 }
 
@@ -195,7 +209,7 @@ func (e *Engine) ClearDistressedParties(ctx context.Context, idgen IDGen, closed
 	for _, cp := range closed {
 		e.pos.open += cp.Size()
 		// get the orders and trades so we can send events to update the datanode
-		o1, o2, t := e.getOrdersAndTrade(cp, idgen, now, mp, mmp)
+		o1, o2, t := e.getOrdersAndTrade(ctx, cp, idgen, now, mp, mmp)
 		orders = append(orders, events.NewOrderEvent(ctx, o1), events.NewOrderEvent(ctx, o2))
 		trades = append(trades, events.NewTradeEvent(ctx, *t))
 		netTrades = append(netTrades, t)
@@ -239,7 +253,7 @@ func (e *Engine) UpdateNetworkPosition(trades []*types.Trade) {
 	}
 }
 
-func (e *Engine) getOrdersAndTrade(pos events.Margin, idgen IDGen, now time.Time, price, dpPrice *num.Uint) (*types.Order, *types.Order, *types.Trade) {
+func (e *Engine) getOrdersAndTrade(ctx context.Context, pos events.Margin, idgen IDGen, now time.Time, price, dpPrice *num.Uint) (*types.Order, *types.Order, *types.Trade) {
 	tSide, nSide := types.SideSell, types.SideBuy // one of them will have to sell
 	s := pos.Size()
 	size := uint64(s)
@@ -261,14 +275,16 @@ func (e *Engine) getOrdersAndTrade(pos events.Margin, idgen IDGen, now time.Time
 		TimeInForce:   types.OrderTimeInForceFOK, // this is an all-or-nothing order, so TIME_IN_FORCE == FOK
 		Type:          types.OrderTypeNetwork,
 		Size:          size,
-		Remaining:     0,
+		Remaining:     size,
 		Side:          nSide,
 	}
+	e.position.RegisterOrder(ctx, &order)
+	order.Remaining = 0
 	partyOrder := types.Order{
 		ID:            idgen.NextID(),
 		MarketID:      e.mID,
 		Size:          size,
-		Remaining:     0,
+		Remaining:     size,
 		Status:        types.OrderStatusFilled,
 		Party:         pos.Party(),
 		Side:          tSide, // assume sell, price is zero in that case anyway
@@ -279,6 +295,8 @@ func (e *Engine) getOrdersAndTrade(pos events.Margin, idgen IDGen, now time.Time
 		TimeInForce:   types.OrderTimeInForceFOK, // this is an all-or-nothing order, so TIME_IN_FORCE == FOK
 		Type:          types.OrderTypeNetwork,
 	}
+	e.position.RegisterOrder(ctx, &partyOrder)
+	partyOrder.Remaining = 0
 	buyParty = order.Party
 	sellParty = partyOrder.Party
 	sellID = partyOrder.ID
@@ -296,12 +314,17 @@ func (e *Engine) getOrdersAndTrade(pos events.Margin, idgen IDGen, now time.Time
 		Aggressor:   order.Side, // we consider network to be aggressor
 		BuyOrder:    buyID,
 		SellOrder:   sellID,
-		Buyer:       buyParty,
-		Seller:      sellParty,
+		Buyer:       types.NetworkParty,
+		Seller:      types.NetworkParty,
 		Timestamp:   now.UnixNano(),
 		Type:        types.TradeTypeNetworkCloseOutBad,
 		SellerFee:   types.NewFee(),
 		BuyerFee:    types.NewFee(),
 	}
+	// settlement engine should see this as a wash trade
+	e.settle.AddTrade(&trade)
+	// the for the rest of the core, this should not seem like a wash trade though...
+	trade.Buyer, trade.Seller = buyParty, sellParty
+	e.position.Update(ctx, &trade, &order, &partyOrder)
 	return &order, &partyOrder, &trade
 }