+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - [9231](https://github.com/vegaprotocol/vega/issues/9231) - Add a `JoinTeam API`
 - [9981](https://github.com/vegaprotocol/vega/issues/9981) - Support filtering on epoch range on transfers.
 - [9981](https://github.com/vegaprotocol/vega/issues/9981) - Support filtering on status on transfers.
+- [10104](https://github.com/vegaprotocol/vega/issues/10104) - Add network position tracking.
 
 ### 🐛 Fixes